requirements.txt
aiofiles
aiohttp
yt-dlp
pyromod
TgCrypto==1.2.5
pyrogram
xmltodict 
requests 
pycryptodomex
